requirements.txt
fastapi>=0.115.0
openai>=1.0.0
openenv-core[core]>=0.2.1
pydantic>=2.12.0
requests>=2.31.0
uvicorn>=0.30.0